Tweaks to run in a container
Some scripts assume that some databases (such as test_slice_##) will be available on localhost.
This won't work since all databases will be hosted in another container (or possible on another host machine), so some changes were needed to allow a configured value for slice_host as an alternative to localhsot.